+++ delocate/cmd/delocate_path.py
@@ -52,5 +52,5 @@ def main() -> None:  # noqa: D103
         )
 
 
-if __name__ == "__main__":
+if __name__ == "__main__":  # pragma: no cover
     main()