+++ delocate/libsana.py
@@ -8,7 +8,6 @@ from __future__ import annotations
 import logging
 import os
 import sys
-import warnings
 from collections.abc import Iterable, Iterator
 from os import PathLike
 from os.path import join as pjoin
@@ -17,8 +16,6 @@ from pathlib import Path
 from typing import (
     Callable,
 )
-
-from typing_extensions import deprecated
 
 from .tmpdirs import TemporaryDirectory
 from .tools import (
@@ -394,83 +391,6 @@ def tree_libs_from_directory(
     )
 
 
-def _allow_all(path: str) -> bool:
-    """Return True for all files."""
-    return True
-
-
-@deprecated("tree_libs doesn't support @loader_path and has been deprecated")
-def tree_libs(
-    start_path: str,
-    filt_func: Callable[[str], bool] | None = None,
-) -> dict[str, dict[str, str]]:
-    """Return analysis of library dependencies within `start_path`.
-
-    Parameters
-    ----------
-    start_path : str
-        root path of tree to search for libraries depending on other libraries.
-    filt_func : None or callable, optional
-        If None, inspect all files for library dependencies. If callable,
-        accepts filename as argument, returns True if we should inspect the
-        file, False otherwise.
-
-    Returns
-    -------
-    lib_dict : dict
-        dictionary with (key, value) pairs of (``libpath``,
-        ``dependings_dict``).
-
-        ``libpath`` is a canonical (``os.path.realpath``) filename of library,
-        or library name starting with {'@loader_path'}.
-
-
-        ``dependings_dict`` is a dict with (key, value) pairs of
-        (``depending_libpath``, ``install_name``), where ``dependings_libpath``
-        is the canonical (``os.path.realpath``) filename of the library
-        depending on ``libpath``, and ``install_name`` is the "install_name" by
-        which ``depending_libpath`` refers to ``libpath``.
-
-    Notes
-    -----
-    See:
-
-    * https://developer.apple.com/library/mac/documentation/Darwin/Reference/ManPages/man1/dyld.1.html
-    * http://matthew-brett.github.io/pydagogue/mac_runtime_link.html
-
-    .. deprecated:: 0.9
-        This function does not support `@loader_path` and only returns the
-        direct dependencies of the libraries in `start_path`.
-
-        :func:`tree_libs_from_directory` should be used instead.
-    """
-    if filt_func is None:
-        filt_func = _allow_all
-    lib_dict: dict[str, dict[str, str]] = {}
-    for dirpath, dirnames, basenames in os.walk(start_path):
-        for base in basenames:
-            depending_path = realpath(pjoin(dirpath, base))
-            for dependency_path, install_name in get_dependencies(
-                depending_path,
-                filt_func=filt_func,
-            ):
-                if dependency_path is None:
-                    # Mimic deprecated behavior.
-                    # A lib_dict with unresolved paths is unsuitable for
-                    # delocating, this is a missing dependency.
-                    dependency_path = realpath(install_name)
-                if install_name.startswith("@loader_path/"):
-                    # Support for `@loader_path` would break existing callers.
-                    logger.debug(
-                        "Excluding %s because it has '@loader_path'.",
-                        install_name,
-                    )
-                    continue
-                lib_dict.setdefault(dependency_path, {})
-                lib_dict[dependency_path][depending_path] = install_name
-    return lib_dict
-
-
 _default_paths_to_search = ("/usr/local/lib", "/usr/lib")
 
 
@@ -553,52 +473,6 @@ def resolve_dynamic_paths(
             return str(abs_path)
 
     raise DependencyNotFound(lib_path)
-
-
-@deprecated(
-    "This function doesn't support @loader_path "
-    "and was replaced by resolve_dynamic_paths"
-)
-def resolve_rpath(lib_path: str, rpaths: Iterable[str]) -> str:
-    """Return `lib_path` with its `@rpath` resolved.
-
-    If the `lib_path` doesn't have `@rpath` then it's returned as is.
-    If `lib_path` has `@rpath` then returns the first `rpaths`/`lib_path`
-    combination found.  If the library can't be found in `rpaths` then a
-    detailed warning is printed and `lib_path` is returned as is.
-
-    Parameters
-    ----------
-    lib_path : str
-        The path to a library file, which may or may not start with `@rpath`.
-    rpaths : sequence of str
-        A sequence of search paths, usually gotten from a call to `get_rpaths`.
-
-    Returns
-    -------
-    lib_path : str
-        A str with the resolved libraries realpath.
-
-    .. deprecated:: 0.9
-        This function does not support `@loader_path`.
-        Use `resolve_dynamic_paths` instead.
-    """
-    if not lib_path.startswith("@rpath/"):
-        return lib_path
-
-    lib_rpath = lib_path.split("/", 1)[1]
-    for rpath in (*rpaths, *_default_paths_to_search):
-        rpath_lib = realpath(pjoin(rpath, lib_rpath))
-        if os.path.exists(rpath_lib):
-            return rpath_lib
-
-    warnings.warn(
-        "Couldn't find {} on paths:\n\t{}".format(
-            lib_path,
-            "\n\t".join(realpath(path) for path in rpaths),
-        )
-    )
-    return lib_path
 
 
 def search_environment_for_lib(lib_path: str | PathLike[str]) -> str: